+++ src/container/signal.rs
@@ -1,6 +1,6 @@
-use std::io::IoError;
+use std::old_io::IoError;
 use std::os::errno;
-use std::io::process::Process;
+use std::old_io::process::Process;
 
 pub use libc::consts::os::posix88::{SIGTERM, SIGINT, SIGQUIT, EINTR, ECHILD};
 use libc::{c_int, pid_t};
@@ -11,8 +11,8 @@ static WNOHANG: c_int = 1;
 
 #[derive(Show)]
 pub enum Signal {
-    Terminate(isize),  // Actual signal for termination: INT, TERM, QUIT...
-    Child(pid_t, isize),  //  pid and result code
+    Terminate(i32),  // Actual signal for termination: INT, TERM, QUIT...
+    Child(pid_t, i32),  //  pid and result code
 }
 
 #[derive(Default)]
@@ -32,30 +32,31 @@ pub fn block_all() {
     unsafe { block_all_signals() };
 }
 
-fn _convert_status(status: i32) -> isize {
+fn _convert_status(status: i32) -> i32 {
     if status & 0xff == 0 {
-        return ((status & 0xff00) >> 8) as isize;
+        return ((status & 0xff00) >> 8);
     }
-    return (128 + (status & 0x7f)) as isize;  // signal
+    return (128 + (status & 0x7f));  // signal
 }
 
 pub fn check_children() -> Option<Signal> {
     let mut status: i32 = 0;
     let pid = unsafe { waitpid(-1, &mut status, WNOHANG) };
+
     if pid > 0 {
         return Some(Child(pid, _convert_status(status)));
     }
     return None
 }
 
-pub fn wait_process(pid: pid_t) -> Result<isize, IoError> {
+pub fn wait_process(pid: pid_t) -> Result<i32, IoError> {
     let mut status: i32 = 0;
     loop {
         let pid = unsafe { waitpid(pid, &mut status, 0) };
         if pid > 0 {
             return Ok(_convert_status(status));
         }
-        if errno() == EINTR as usize {
+        if errno() == EINTR {
             continue;
         }
         return Err(IoError::last_error());
@@ -63,6 +64,6 @@ pub fn wait_process(pid: pid_t) -> Result<isize, IoError> {
 }
 
 
-pub fn send_signal(pid: pid_t, sig: isize) {
-    Process::kill(pid, sig).ok();
+pub fn send_signal(pid: pid_t, sig: i32) {
+    Process::kill(pid, sig as isize).ok();
 }
